reports_api/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
pymongo==3.12.1
openpyxl==3.0.9
python-pptx==0.6.19